Maps and Board refactoring
